# [Akka.Actor][Perf] Reduce `Envelope` copying in message-processing hotpaths