--- a/src/benchmark/Akka.Benchmarks/Configurations/Configs.cs
+++ b/src/benchmark/Akka.Benchmarks/Configurations/Configs.cs
@@ -6,7 +6,6 @@
 //-----------------------------------------------------------------------
 
 using System.Reflection;
-using Akka.Benchmarks.Actor;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Columns;
 using BenchmarkDotNet.Configs;