--- a/src/core/Akka.TestKit/Internal/InternalTestActorRef.cs
+++ b/src/core/Akka.TestKit/Internal/InternalTestActorRef.cs
@@ -237,7 +237,7 @@ namespace Akka.TestKit.Internal
             /// TBD
             /// </summary>
             /// <param name="envelope">TBD</param>
-            protected internal override void AutoReceiveMessage(Envelope envelope)
+            protected internal override void AutoReceiveMessage(in Envelope envelope)
             {
                 if (envelope.Message is InternalGetActor)
                     Sender.Tell(Actor, Self);