--- a/src/core/Akka/Dispatch/AbstractDispatcher.cs
+++ b/src/core/Akka/Dispatch/AbstractDispatcher.cs
@@ -682,7 +682,7 @@ channel-executor.priority = normal");
         /// </summary>
         /// <param name="cell">TBD</param>
         /// <param name="envelope">TBD</param>
-        public virtual void Dispatch(ActorCell cell, Envelope envelope)
+        public virtual void Dispatch(ActorCell cell, in Envelope envelope)
         {
             var mbox = cell.Mailbox;
             mbox.Enqueue(cell.Self, envelope);