--- a/src/core/Akka/Dispatch/MessageQueues/BlockingMessageQueue.cs
+++ b/src/core/Akka/Dispatch/MessageQueues/BlockingMessageQueue.cs
@@ -59,7 +59,7 @@ namespace Akka.Dispatch.MessageQueues
         /// </summary>
         /// <param name="receiver">TBD</param>
         /// <param name="envelope">TBD</param>
-        public void Enqueue(IActorRef receiver, Envelope envelope)
+        public void Enqueue(IActorRef receiver, in Envelope envelope)
         {
             lock (_lock)
             {