--- a/src/core/Akka/Dispatch/MessageQueues/UnboundedPriorityMessageQueue.cs
+++ b/src/core/Akka/Dispatch/MessageQueues/UnboundedPriorityMessageQueue.cs
@@ -79,7 +79,7 @@ namespace Akka.Dispatch.MessageQueues
             return false;
         }
 
-        public void EnqueueFirst(Envelope envelope)
+        public void EnqueueFirst(in Envelope envelope)
         {
             _prependBuffer.Push(envelope);
         }