--- a/src/core/Akka/Routing/RoutedActorCell.cs
+++ b/src/core/Akka/Routing/RoutedActorCell.cs
@@ -188,7 +188,7 @@ namespace Akka.Routing
         /// TBD
         /// </summary>
         /// <param name="envelope">TBD</param>
-        public override void SendMessage(Envelope envelope)
+        public override void SendMessage(in Envelope envelope)
         {
             if (RouterConfig.IsManagementMessage(envelope.Message))
                 base.SendMessage(envelope);